hf parent class
