+++ llmfoundry/models/hf/hf_causal_lm.py
@@ -71,6 +71,11 @@ class ComposerHFCausalLM(HuggingFaceModelWithFSDP):
         tokenizer (PreTrainedTokenizer): The tokenizer that the model will use.
     """
 
+    model_cls: Union[_BaseAutoModelClass,
+                     PreTrainedModel] = AutoModelForCausalLM
+    default_train_metrics: list = DEFAULT_CAUSAL_LM_TRAIN_METRICS
+    default_eval_metrics: list = DEFAULT_CAUSAL_LM_EVAL_METRICS
+
     def __init__(
         self,
         tokenizer: PreTrainedTokenizerBase,
@@ -90,319 +95,23 @@ class ComposerHFCausalLM(HuggingFaceModelWithFSDP):
         additional_eval_metrics: Optional[list] = None,
         should_save_peft_only: bool = True,
     ):
-
-        config_overrides = config_overrides or {}
-
-        model = ComposerHFCausalLM.build_inner_model(
-            pretrained_model_name_or_path=pretrained_model_name_or_path,
+        super().__init__(
+            pretrained_model_name_or_path,
+            tokenizer=tokenizer,
+            pretrained=pretrained,
             pretrained_lora_id_or_path=pretrained_lora_id_or_path,
             trust_remote_code=trust_remote_code,
-            init_device=init_device,
-            use_flash_attention_2=use_flash_attention_2,
             use_auth_token=use_auth_token,
-            config_overrides=config_overrides,
+            use_flash_attention_2=use_flash_attention_2,
             load_in_8bit=load_in_8bit,
-            pretrained=pretrained,
-            prepare_for_fsdp=False,
-        )
-
-        model = self.transform_model(model)
-
-        ComposerHFCausalLM.prepare_inner_model(model, init_device)
-
-        train_metrics, eval_metrics = ComposerHFCausalLM.build_metrics(
+            init_device=init_device,
+            config_overrides=config_overrides,
+            use_logits=True,
+            shift_labels=True,
+            peft_config=peft_config,
+            allow_embedding_resizing=allow_embedding_resizing,
             use_train_metrics=use_train_metrics,
             additional_train_metrics=additional_train_metrics,
             additional_eval_metrics=additional_eval_metrics,
-        )
-
-        if peft_config is not None and not peft_installed:
-            raise ValueError(
-                'PEFT is not installed, but peft_config was passed. Please install LLM Foundry with the peft extra to use peft_config.',
-            )
-
-        peft_config_object = None
-        if peft_config is not None:
-            peft_config_object = self.get_peft_config(peft_config)
-
-        # Set up config args for the model construction and base classes
-        super().__init__(
-            model=model,
-            shift_labels=True,
-            tokenizer=tokenizer,
-            metrics=train_metrics,
-            eval_metrics=eval_metrics,
-            allow_embedding_resizing=allow_embedding_resizing,
-            init_device=init_device,
-            peft_config=peft_config_object,
             should_save_peft_only=should_save_peft_only,
         )
-
-    def transform_model(self, model: PreTrainedModel) -> PreTrainedModel:
-        """Transforms the model after initialization.
-
-        Args:
-            model (PreTrainedModel): The model to transform.
-
-        Returns:
-            PreTrainedModel: The transformed model.
-        """
-        return model
-
-    @staticmethod
-    def build_metrics(
-        use_train_metrics: bool,
-        additional_train_metrics: Optional[list[str]] = None,
-        additional_eval_metrics: Optional[list[str]] = None,
-    ) -> tuple[list[Metric], list[Metric]]:
-        """Builds the training and evaluation metrics for the model.
-
-        Args:
-            use_train_metrics (bool): Whether to use training metrics.
-            additional_train_metrics (Optional[List[str]]): Additional training metrics to include.
-            additional_eval_metrics (Optional[List[str]]): Additional evaluation metrics to include.
-
-        Returns:
-            Tuple[List[Metric], List[Metric]]: A tuple containing the list of training metrics and evaluation metrics.
-        """
-        from llmfoundry.utils.builders import build_metric
-
-        train_metric_names = DEFAULT_CAUSAL_LM_TRAIN_METRICS + (
-            additional_train_metrics or []
-        )
-        train_metrics = [
-            build_metric(metric, {}) for metric in train_metric_names
-        ] if use_train_metrics else []
-        eval_metric_names = DEFAULT_CAUSAL_LM_EVAL_METRICS + (
-            additional_eval_metrics or []
-        )
-        eval_metrics = [
-            build_metric(metric, {}) for metric in eval_metric_names
-        ]
-
-        return train_metrics, eval_metrics
-
-    @staticmethod
-    def build_inner_model(
-        pretrained_model_name_or_path: str,
-        pretrained_lora_id_or_path: Optional[str],
-        trust_remote_code: bool,
-        init_device: str,
-        use_flash_attention_2: bool,
-        use_auth_token: bool,
-        config_overrides: dict[str, Any],
-        load_in_8bit: bool,
-        pretrained: bool,
-        model_cls: Union[_BaseAutoModelClass,
-                         PreTrainedModel] = AutoModelForCausalLM,
-        prepare_for_fsdp: bool = False,
-    ) -> Union[PreTrainedModel, 'PeftModel']:
-        """Builds the inner model for the ComposerHFCausalLM.
-
-        Args:
-            pretrained_model_name_or_path (str): The pretrained model name or path.
-            pretrained_lora_id_or_path (Optional[str]): The pretrained LORA ID or path.
-            trust_remote_code (bool): Whether to trust remote code.
-            init_device (str): The initialization device.
-            use_flash_attention_2 (bool): Whether to use flash attention 2.
-            use_auth_token (bool): Whether to use an authentication token.
-            config_overrides (Dict[str, Any]): The configuration overrides.
-            load_in_8bit (bool): Whether to load in 8-bit.
-            pretrained (bool): Whether the model is pretrained.
-            model_cls (Union[Type, Type[PreTrainedModel]]): HF class for models. Default: ``AutoModelForCausalLM``.
-            prepare_for_fsdp (bool, optional): Whether to prepare the model for FSDP wrapping. Default: ``False``.
-
-        Returns:
-            Union[PreTrainedModel, 'PeftModel']: The built inner model.
-            prepare_for_fsdp (bool): Whether to prepare the model for FSDP wrapping. Default: ``False``.
-        """
-        if not trust_remote_code and pretrained_model_name_or_path.startswith(
-            'mosaicml/mpt',
-        ):
-            raise ValueError(
-                'trust_remote_code must be set to True for MPT models. Without this, the MPT model code will come from the transformers library, '
-                +
-                'which is significantly slower and not compatible with the LLM foundry training code, rather than the code release by MosaicML.',
-            )
-        # Resolve "mixed" init device to either "cpu" or "meta"
-        resolved_init_device = hf_get_init_device(init_device)
-        requested_attention_implementation = 'flash_attention_2' if use_flash_attention_2 else 'eager'
-
-        if use_flash_attention_2 and not is_flash_v2_installed():
-            raise ValueError(
-                'use_flash_attention_2 is set to True, but flash-attention 2 is not installed. '
-                + 'Please `pip install llm-foundry[gpu]`.',
-            )
-
-        if not (
-            hasattr(model_cls, 'from_pretrained') and
-            hasattr(model_cls, 'from_config')
-        ):
-            raise AttributeError(
-                f'{model_cls=} is missing `from_pretrained` and `from_config` support.',
-            )
-
-        # Hugging Face copies the modules into the
-        # transformers modules cache. On particular systems, this operation seems to cause contention between
-        # the different processes. To avoid this contention, we first create the config and generation config on local rank
-        # zero. This will set up the transformers module cache and avoid the future contention.
-        if dist.get_local_rank() == 0:
-            AutoConfig.from_pretrained(
-                pretrained_model_name_or_path,
-                trust_remote_code=trust_remote_code,
-                use_auth_token=use_auth_token,
-                attn_implementation=requested_attention_implementation,
-                use_cache=
-                False,  # Necessary due to https://github.com/huggingface/transformers/issues/28056
-            )
-            try:
-                GenerationConfig.from_pretrained(
-                    pretrained_model_name_or_path,
-                    use_auth_token=use_auth_token,
-                )
-            except OSError:
-                pass
-
-        dist.barrier()
-
-        # Construct the Hugging Face config to use
-        config = AutoConfig.from_pretrained(
-            pretrained_model_name_or_path,
-            trust_remote_code=trust_remote_code,
-            use_auth_token=use_auth_token,
-            attn_implementation=requested_attention_implementation,
-            use_cache=
-            False,  # Necessary due to https://github.com/huggingface/transformers/issues/28056
-        )
-
-        set_config_overrides(config, config_overrides)
-
-        # We need to have all non-zero local ranks be not-pretrained
-        # Rank 0 will still be pretrained, and distribute the weights appropriately
-        if dist.get_local_rank() != 0 and init_device == 'mixed':
-            pretrained = False
-
-        # Hugging Face copies the modules into the
-        # transformers modules cache. On particular systems, this operation seems to cause contention between
-        # the different processes. To avoid this contention, we first create the model (on meta device) on local rank
-        # zero. This will set up the transformers model cache and avoid the future contention.
-        if dist.get_local_rank() == 0:
-            if pretrained and os.path.isdir(pretrained_model_name_or_path):
-                with init_empty_weights(include_buffers=False):
-                    with warnings.catch_warnings():
-                        warnings.simplefilter('ignore', UserWarning)
-                        model_cls.from_pretrained(
-                            pretrained_model_name_or_path,
-                            trust_remote_code=trust_remote_code,
-                            use_auth_token=use_auth_token,
-                            attn_implementation=
-                            requested_attention_implementation,
-                            config=config,
-                        )
-            else:
-                with init_empty_weights(include_buffers=False):
-                    model_cls.from_config(
-                        config,
-                        trust_remote_code=trust_remote_code,
-                        attn_implementation=requested_attention_implementation,
-                    )
-
-        dist.barrier()
-
-        # initialize the model on the correct device
-        if resolved_init_device == 'cpu':
-            if pretrained:
-                model = model_cls.from_pretrained(
-                    pretrained_model_name_or_path,
-                    trust_remote_code=trust_remote_code,
-                    use_auth_token=use_auth_token,
-                    load_in_8bit=load_in_8bit,
-                    attn_implementation=requested_attention_implementation,
-                    config=config,
-                )
-            else:
-                model = model_cls.from_config(
-                    config,
-                    trust_remote_code=trust_remote_code,
-                    attn_implementation=requested_attention_implementation,
-                )
-        elif resolved_init_device == 'meta':
-            if pretrained:
-                raise ValueError(
-                    'Setting cfg.pretrained=True is not supported when init_device="meta".',
-                )
-            with init_empty_weights(include_buffers=False):
-                model = model_cls.from_config(
-                    config,
-                    trust_remote_code=trust_remote_code,
-                    attn_implementation=requested_attention_implementation,
-                )
-        else:
-            raise ValueError(
-                f'init_device="{init_device}" must be either "cpu" or "meta".',
-            )
-
-        signal_file_path = f'.node_{dist.get_node_rank()}_local_rank0_completed'
-        if dist.get_local_rank() == 0:
-            with open(signal_file_path, 'wb') as f:
-                f.write(b'local_rank0_completed_download')
-
-        # Avoid the collective call until the local rank zero has finished trying to download the checkpoint
-        # so that we don't timeout for large downloads. This syncs all processes on the node
-        with dist.local_rank_zero_download_and_wait(signal_file_path):
-            # Then, wait to ensure every node has finished downloading the checkpoint
-            dist.barrier()
-
-        if dist.get_local_rank() == 0:
-            os.remove(signal_file_path)
-
-        # Use the pretrained generation config for the model if it exists.
-        try:
-            model.generation_config = GenerationConfig.from_pretrained(
-                pretrained_model_name_or_path,
-                use_auth_token=use_auth_token,
-            )
-        except OSError:
-            log.warning(
-                f'No existing generation config found for the model with name or path {pretrained_model_name_or_path}. Using default generation config.',
-            )
-
-        # Hugging Face's weight tying does not succeed if the model is inited on meta device
-        # so we manually apply the weight tying here
-        if model.config.tie_word_embeddings and resolved_init_device == 'meta':
-            model.tie_weights()
-
-        if pretrained_lora_id_or_path is not None:
-            if not peft_installed:
-                raise ValueError(
-                    'PEFT is not installed, but lora_id_or_path was passed. Please install LLM Foundry with the peft extra to use lora_id_or_path.',
-                )
-            from peft import PeftModelForCausalLM
-            model = PeftModelForCausalLM.from_pretrained(
-                model,
-                pretrained_lora_id_or_path,
-            )
-
-        if prepare_for_fsdp:
-            ComposerHFCausalLM.prepare_inner_model(model, init_device)
-
-        return model
-
-    def get_peft_config(self, peft_config_dict: dict[str, Any]) -> 'PeftConfig':
-        if peft_installed:
-            from peft import LoraConfig
-            peft_type = peft_config_dict.get('peft_type', '')
-            if peft_type.upper() != 'LORA':
-                raise ValueError(
-                    f'Only LORA is supported for peft_type, but got {peft_type}.',
-                )
-            task_type = peft_config_dict.get('task_type', '')
-            if task_type.upper() != 'CAUSAL_LM':
-                raise ValueError(
-                    f'Only CAUSAL_LM is supported for task_type, but got {task_type}.',
-                )
-            return LoraConfig(**peft_config_dict)
-        else:
-            raise ValueError(
-                'PEFT is not installed, but peft_config was passed. Please install LLM Foundry with the peft extra to use peft_config.',
-            )
